requirements.txt
python-dotenv
langchain-openai
langchain-community
PyGithub
python-slugify
schedule
